requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements/prod.in
#
aiohttp==3.8.1
    # via gradio
aiosignal==1.2.0
    # via aiohttp
analytics-python==1.4.0
    # via gradio
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
asgiref==3.5.2
    # via uvicorn
async-timeout==4.0.2
    # via aiohttp
asynctest==0.13.0
    # via aiohttp
attrs==21.4.0
    # via aiohttp
backoff==1.10.0
    # via analytics-python
backports.zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
bcrypt==3.2.2
    # via paramiko
boto3==1.24.40
    # via
    #   boto3-extensions
    #   smart-open
boto3-extensions==0.19.0
    # via gantry
botocore==1.27.40
    # via
    #   boto3
    #   boto3-extensions
    #   s3transfer
cached-property==1.5.2
    # via h5py
cachetools==4.2.4
    # via gantry
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
click==8.1.2
    # via
    #   gantry
    #   uvicorn
click-spinner==0.1.10
    # via gantry
colorama==0.4.5
    # via
    #   gantry
    #   halo
    #   log-symbols
cryptography==37.0.2
    # via paramiko
cycler==0.11.0
    # via matplotlib
dateparser==1.1.1
    # via gantry
fastapi==0.78.0
    # via gradio
ffmpy==0.3.0
    # via gradio
fonttools==4.33.3
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2022.5.0
    # via gradio
gantry==0.4.9
    # via -r requirements/prod.in
gradio==3.0.21
    # via -r requirements/prod.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
h5py==3.6.0
    # via -r requirements/prod.in
halo==0.0.31
    # via gantry
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via gradio
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
importlib-metadata==4.11.3
    # via
    #   -r requirements/prod.in
    #   click
isodate==0.6.1
    # via gantry
jinja2==2.11.3
    # via
    #   -r requirements/prod.in
    #   gradio
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kiwisolver==1.4.2
    # via matplotlib
linkify-it-py==1.0.3
    # via markdown-it-py
log-symbols==0.0.14
    # via halo
markdown-it-py[linkify,plugins]==2.1.0
    # via
    #   gradio
    #   mdit-py-plugins
markupsafe==2.0.1
    # via
    #   -r requirements/prod.in
    #   jinja2
marshmallow==3.17.0
    # via
    #   gantry
    #   marshmallow-oneofschema
marshmallow-oneofschema==3.0.1
    # via gantry
matplotlib==3.5.2
    # via gradio
mdit-py-plugins==0.3.0
    # via markdown-it-py
mdurl==0.1.1
    # via markdown-it-py
monotonic==1.6
    # via
    #   analytics-python
    #   gantry
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.21.6
    # via
    #   -r requirements/prod.in
    #   gantry
    #   gradio
    #   h5py
    #   matplotlib
    #   pandas
    #   torchvision
orjson==3.7.2
    # via gradio
packaging==21.3
    # via
    #   marshmallow
    #   matplotlib
pandas==1.3.5
    # via
    #   gantry
    #   gradio
paramiko==2.11.0
    # via gradio
pillow==7.1.2
    # via
    #   -r requirements/prod.in
    #   gradio
    #   matplotlib
    #   torchvision
pycparser==2.21
    # via cffi
pycryptodome==3.14.1
    # via gradio
pydantic==1.9.1
    # via
    #   fastapi
    #   gradio
pydub==0.25.1
    # via gradio
pynacl==1.5.0
    # via paramiko
pyngrok==5.1.0
    # via -r requirements/prod.in
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.1
    # via
    #   analytics-python
    #   botocore
    #   dateparser
    #   gantry
    #   matplotlib
    #   pandas
python-multipart==0.0.5
    # via gradio
pytz==2022.1
    # via
    #   dateparser
    #   pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   gantry
    #   pyngrok
regex==2022.3.2
    # via dateparser
requests==2.27.1
    # via
    #   -r requirements/prod.in
    #   analytics-python
    #   gantry
    #   gradio
    #   torchvision
rfc3986[idna2008]==1.5.0
    # via httpx
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   analytics-python
    #   halo
    #   isodate
    #   paramiko
    #   python-dateutil
    #   python-multipart
smart-open[s3]==5.2.1
    # via -r requirements/prod.in
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
spinners==0.0.24
    # via halo
starlette==0.19.1
    # via fastapi
tabulate==0.8.10
    # via gantry
termcolor==1.1.0
    # via halo
torch==1.12.0
    # via
    #   -r requirements/prod.in
    #   torchvision
torchvision==0.13.0
    # via -r requirements/prod.in
tqdm==4.64.0
    # via
    #   -r requirements/prod.in
    #   gantry
typeguard==2.13.3
    # via gantry
typing-extensions==4.1.1
    # via
    #   aiohttp
    #   anyio
    #   asgiref
    #   async-timeout
    #   gantry
    #   importlib-metadata
    #   kiwisolver
    #   markdown-it-py
    #   pydantic
    #   starlette
    #   torch
    #   torchvision
    #   uvicorn
    #   yarl
tzdata==2022.1
    # via pytz-deprecation-shim
tzlocal==4.2
    # via dateparser
uc-micro-py==1.0.1
    # via linkify-it-py
urllib3==1.26.9
    # via
    #   botocore
    #   requests
uvicorn==0.17.6
    # via gradio
yarl==1.7.2
    # via aiohttp
zipp==3.8.0
    # via importlib-metadata
